Fix accidentally drawing the clipping path in O2PDFContext
The "W" operator set the clipping path to the current path, but it does not
consume the current path; which means we were continuing to append to the
same path when we thought we started the next one. Fix this by emitting the
"n" (path no-op) operator, which seems to be the usual pattern for doing this.