Make sure to set `_viewsNeedDisplay` when we're told `NO`
This fixes rendering in the OpenGL sample app.

What was happening was that on the initial call to `setViewsNeedDisplay` with `YES`, we setup a performer and set `_viewsNeedDisplay`. When the performer fires, it redraws and then calls `setViewsNeedDisplay` with `NO`. However, because our conditional was checking if the argument was `NO` and would return early in that case, we never set `_viewsNeedDisplay` to `NO`, and thus future calls to `setViewsNeedDisplay` saw `_viewsNeedDisplay` was still `YES` and returned early without setting up a performer.