Workaround freezing bug on window close requests
I'm not entirely sure what was wrong with the old approach of calling it directly, but scheduling the call onto the run loop fixes the freezing.

This commit also fixes a segfault when accessing an invalid window via `lastFocusedWindow`. The issue is that we sometimes don't get FocusOut before a window dies, so we have to set it to nil when it tells us that it's dying.